% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.debugger.R
\docType{data}
\name{allKeywords}
\alias{allKeywords}
\title{Standard Quantlet keywords inherited from the older PHP based QuantNet.}
\format{A character vector with 594 keywords}
\source{
\url{http://quantlet.de/}
}
\usage{
allKeywords
}
\description{
A character vector containing 594 standard Quantlet keywords.
Please use as many as possible of them to enable good text mining results for the D3 Visualization.
}
\keyword{datasets}

