% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.books.R
\name{matchBook}
\alias{matchBook}
\title{Postprocessing for the data field \code{Published in} from the parsed YAML object.
NULL values are corrected to empty string. Further, a matching against the shortcut list is performed.}
\usage{
matchBook(publ_v)
}
\arguments{
\item{publ_v}{data field \code{Published in} from the parsed YAML object}
}
\value{
The matched shortcut if such one is available in the \code{book_match_l} list otherwise the original string
}
\description{
Postprocessing for the data field \code{Published in} from the parsed YAML object.
NULL values are corrected to empty string. Further, a matching against the shortcut list is performed.
}
\examples{
matchBook("statistics of financial markets i")
}

