% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.Qdfields.R
\docType{data}
\name{Q_dfield_list}
\alias{Q_dfield_list}
\title{YAML data field matching list. Used by \code{yaml.getQField}, \code{yaml.not.Qdfields}, \code{yaml.Qdfields.from.meta} and \code{yaml.Qdfields.nchar.from.meta}.}
\format{An object of class \code{list} of length 17.}
\usage{
Q_dfield_list
}
\description{
YAML data field matching list. Used by \code{yaml.getQField}, \code{yaml.not.Qdfields}, \code{yaml.Qdfields.from.meta} and \code{yaml.Qdfields.nchar.from.meta}.
}
\keyword{datasets}

