% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamldebugger.R
\name{yaml.debugger.get.qnames}
\alias{yaml.debugger.get.qnames}
\title{Loads the list of Q folders. RootPath is required where the Qs are located.}
\usage{
yaml.debugger.get.qnames(RootPath, show_qnames = TRUE)
}
\arguments{
\item{RootPath}{current working directory with Qs for debugging}

\item{show_qnames}{boolean trigger for showing / not showing the found Q names in the folder RootPath}
}
\value{
the list with Q names needed in the further process.
}
\description{
Loads the list of Q folders. RootPath is required where the Qs are located.
}
\examples{
qnames = yaml.debugger.get.qnames(d_init$RootPath)
}

