% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.debugger.R
\name{yaml.debugger.run}
\alias{yaml.debugger.run}
\title{Main part of the YAML debugger. Loops trough the provided Q folders and extracts YAML meta info.
If errors occur, corresponding error handling is executed.}
\usage{
yaml.debugger.run(qnames, init.obj)
}
\arguments{
\item{qnames}{the list with Q folder names as provided by \code{yaml.debugger.get.qnames}}

\item{init.obj}{as provided by \code{yaml.debugger.init}}
}
\value{
The summary data of the YAML debugger process. Possible YAML parser errors are also provided in this data structure
}
\description{
Main part of the YAML debugger. Loops trough the provided Q folders and extracts YAML meta info.
If errors occur, corresponding error handling is executed.
}
\examples{
\dontrun{
d_results = yaml.debugger.run(qnames, d_init)
}
}

