% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamldebugger.R
\name{yaml.debugger.summary}
\alias{yaml.debugger.summary}
\title{Overview of the parser results. If errors occured, the corresponding Q names and YAML errors are displayed.}
\usage{
yaml.debugger.summary(qnames, results, showErrors = TRUE,
  showOverView = TRUE, summaryType = "full")
}
\arguments{
\item{qnames}{the list with Q names as provided by 'yaml.debugger.get.qnames'}

\item{results}{as provided by 'yaml.debugger.run'}

\item{showErrors}{boolean trigger for showing / not showing the parser errors}

\item{showOverView}{boolean trigger for showing / not showing the total Overview}

\item{summaryType}{[mini/compact/full] controls the details and extent of the Overview output}
}
\value{
the summary overview in matrix form for further inspection
}
\description{
Overview of the parser results. If errors occured, the corresponding Q names and YAML errors are displayed.
}
\examples{
OverView = yaml.debugger.summary(qnames, d_results, summaryType = "mini")
}

