% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.Qdfields.R
\name{yaml.getQField}
\alias{yaml.getQField}
\title{Extracts the desired data field from the parsed YAML object.}
\usage{
yaml.getQField(yaml_dset, field_name)
}
\arguments{
\item{yaml_dset}{parsed YAML object as a structured list}

\item{field_name}{shortcut for the desired data field for extraction as defined in \code{Q_dfield_list}}
}
\value{
A collapsed character vector containing all text values of the corresponding full name variations of
the given data field shortcut.
}
\description{
\code{Q_dfield_list} is the list defining the data field shortcuts and the corresponding full name variations.
}
\examples{
\dontrun{
yaml.getQField(yaml_meta, "q")
}
}

