% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamldebugger.R
\name{yaml.getQName}
\alias{yaml.getQName}
\title{YAML postprocessing allowing fault tolerance.}
\usage{
yaml.getQName(yaml_dset)
}
\arguments{
\item{yaml_dset}{YAML data object}
}
\value{
the content of the data field which corresponds to the "Quantlet name"
}
\description{
YAML postprocessing allowing fault tolerance.
}

