% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.Qdfields.R
\name{yaml.not.Qdfields}
\alias{yaml.not.Qdfields}
\title{Shows the YAML data fields (of a parsed repository with Quantlet folders) which are not captured by \code{Q_dfield_list} so far.}
\usage{
yaml.not.Qdfields(meta_names_distribution)
}
\arguments{
\item{meta_names_distribution}{table of disjoint YAML data fields which were extracted by \code{yaml.debugger.run} from the
given repository with Quantlet folders}
}
\value{
A character vector with YAML data fields which are not captured by \code{Q_dfield_list} so far.
}
\description{
\code{Q_dfield_list} is the list defining the data field shortcuts and the corresponding full name variations.
}
\examples{
\dontrun{
d_results = yaml.debugger.run(qnames, d_init)
yaml.not.Qdfields(d_results$meta_names_distribution)
}
}

