% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.Qdfields.R
\name{yaml.Qdfields.nchar.from.meta}
\alias{yaml.Qdfields.nchar.from.meta}
\title{Counts the number of characters in the YAML data field shortcuts as defined in \code{Q_dfield_list} for a given parsed YAML object.}
\usage{
yaml.Qdfields.nchar.from.meta(yaml_meta)
}
\arguments{
\item{yaml_meta}{parsed YAML object as a structured list}
}
\value{
A named vector with the total number of characters in each found data field shortcut for a given parsed YAML object.
}
\description{
\code{Q_dfield_list} is the list defining the data field shortcuts and the corresponding full name variations.
}
\examples{
\dontrun{
d_results = yaml.debugger.run(qnames, d_init)
rowSums(sapply( d_results$Metainfos, function(yaml){ yaml.Qdfields.nchar.from.meta(yaml) } ))
}
}

