% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.debugger.R
\name{yaml.debugger.init}
\alias{yaml.debugger.init}
\title{YAML initialization function.}
\usage{
yaml.debugger.init(RootPath = "C:/r/test", show_keywords = FALSE)
}
\arguments{
\item{RootPath}{current working directory with Qs for debugging}

\item{show_keywords}{boolean trigger for showing / not showing the standard keywords}
}
\value{
The list with data needed in the further process.
}
\description{
YAML initialization function.
}
\examples{
\dontrun{
d_init = yaml.debugger.init("c:/test", show_keywords = TRUE)
}
}

