% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.debugger.R
\name{yaml.debugger.summary}
\alias{yaml.debugger.summary}
\title{Overview of the parser results. If errors occured, the corresponding Q folder names and YAML errors are displayed.}
\usage{
yaml.debugger.summary(qfolders, results, showErrors = TRUE,
  showOverView = TRUE, summaryType = "full")
}
\arguments{
\item{qfolders}{the list with Q folder names as provided by \code{yaml.debugger.get.qnames}}

\item{results}{as provided by \code{yaml.debugger.run}}

\item{showErrors}{boolean trigger for showing / not showing the parser errors}

\item{showOverView}{boolean trigger for showing / not showing the total \code{Overview}}

\item{summaryType}{[mini/compact/full] controls the details and extent of the \code{Overview} output}
}
\value{
The summary overview as data frame for further inspection
}
\description{
Overview of the parser results. If errors occured, the corresponding Q folder names and YAML errors are displayed.
}
\examples{
\dontrun{
OverView = yaml.debugger.summary(qfolders, d_results, summaryType = "mini")
}
}

